+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# filepath: /C:/Users/Gabe/youtube_transcript_dart/CHANGELOG.md
+## 0.1.0
+
+- Initial release of transcriptify.
+- Added support for fetching YouTube video transcripts.
+
 ## 1.0.0
 
 - Initial version.